Testing: Move all mocked files to mocks subfolders
