+++ client/lib/analytics/test/mocks/config/index.js
