+++ client/lib/analytics/test/mocks/lib/load-script/index.js
