+++ client/lib/feed-post-store/test/mocks/lib/post-normalizer.js
