+++ client/lib/feed-post-store/test/mocks/lib/wp.js
