+++ client/lib/feed-stream-store/test/mocks/lib/data-poller.js
