+++ client/lib/feed-stream-store/test/mocks/lib/post-normalizer.js
