+++ client/lib/feed-stream-store/test/mocks/lib/wp.js
