+++ client/lib/follow-list/test/index.js
@@ -1,5 +1,5 @@
 /** @jest-environment jsdom */
-jest.mock( 'lib/wp', () => require( './lib/wp' ) );
+jest.mock( 'lib/wp', () => require( './mocks/lib/wp' ) );
 
 var assert = require( 'chai' ).assert,
 	sinon = require( 'sinon' );
