+++ client/lib/like-store/test/index.js
@@ -1,4 +1,4 @@
-jest.mock( 'lib/wp', () => require( './lib/wp' ) );
+jest.mock( 'lib/wp', () => require( './mocks/lib/wp' ) );
 
 /**
  * External Dependencies
