+++ client/lib/like-store/test/mocks/lib/wp.js
