+++ client/lib/post-normalizer/test/mocks/lib/safe-image-url.js
