+++ client/lib/signup/test/dependency-store.js
@@ -1,6 +1,6 @@
 /** @jest-environment jsdom */
 jest.mock( 'lib/user', () => () => {} );
-jest.mock( 'signup/config/steps', () => require( './signup/config/steps' ) );
+jest.mock( 'signup/config/steps', () => require( './mocks/signup/config/steps' ) );
 
 /**
  * External dependencies
