+++ client/lib/signup/test/mocks/signup/config/steps.js
