+++ client/signup/test/mocks/lib/user/index.js
